sensor_Data/requirements.txt
numpy
opencv-python
py-opc-ng
pyusbiss
pyubx2
pyserial